+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * `@PostDispose`
   * `@PreDispose`
 * Remove the `@OnStale` annotation. It has never been used in any downstream library or application.
+* Add the ability to suppress the warning if the `allowEmpty=true` is specified on the `@ArezComponent` annotation but there are reactive components present. Previously this scenario was an error and only suppressable by if the component was also marked as `@Generated`.
 
 ### [v0.156](https://github.com/arez/arez/tree/v0.156) (2019-11-21) · [Full Changelog](https://github.com/arez/arez/compare/v0.155...v0.156) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.155&new=0.156)